requirements.txt
pvporcupine
SpeechRecognition
openai
gtts
audio2numpy
pyaudio
simpleaudio
scipy
lifxlan
xmltodict
textdistance
tensorflow
pillow
